refactor: extrai estúdio, ranking e atividade para services
Conclui a camada de services. O estúdio (módulos, aulas, questões, publicação, editor, exclusões e salvar a aula inteira) vai para estudio.service. O ranking global vira rankingGlobal no ranking.ts e as atividades recentes vão para activity.service. Com isso o TrailController fica só com parse e resposta, caindo de 1452 para 420 linhas. Comportamento da API preservado.
